make mistakes a requireable module, use browservefy for local dev
