+++ README.md
@@ -49,6 +49,10 @@ and drop Javascript files in the `local/` directory. I've included
 Then [boot up a server](https://gist.github.com/tmcw/4989751), and go to
 `http://localhost:3000/#yourfile.js` or whatever.
 
+## Developing Locally
+
+Clone this repo and run `npm install` && `npm start`
+
 ## See Also
 
 * [jsconsole](http://jsconsole.com/)
